test: 166 diagnostics out of discovery; 1778 proxy-env isolation (#2457)
Issue 166 (2 baselined setup errors): both drivers are manual
diagnostics — test_qwen_cli/test_stream_json_mode take env_name/env_mods
as plain main() arguments (pytest read them as missing fixtures) and
shell out to the real qwen CLI, which lane checkouts never have. Moved to
manual_*.py outside discovery with MANUAL-DIAGNOSTIC docstrings (172
pattern); the historical paths keep pointer+guard tests so the lane's
file-based round-robin shard layout is unchanged.

Issue 1778 (2 baselined failures): safe_request deliberately supports
proxies from environment variables (_get_proxies) and requests'
trust_env merges OS-level proxy settings (macOS System Configuration
among them), so any proxy-configured runner saw extra entries merged
into the final mapping. The two contract tests now pin the merge
semantics under a neutralized proxy environment instead of inheriting
the runner's.

Lane reproduce (sqlite HOME): 4 failed pre-change; 23/23 post-change.
